Feature/roles (#3)
implemented role model

+++ api/user_router.py
@@ -12,6 +12,7 @@ from api.actions.user import _create_new_user
 from api.actions.user import _delete_user
 from api.actions.user import _get_user_by_id
 from api.actions.user import _update_user
+from api.actions.user import check_user_permissions
 from api.schemas import UpdateUserRequest
 from api.schemas import UserCreate
 from api.schemas import UserShow
@@ -42,14 +43,21 @@ async def delete_user(
     session: AsyncSession = Depends(get_async_session),
     current_user: User = Depends(get_current_user_from_token),
 ) -> UserShow:
-    user = await _delete_user(user_id, session)
-    if user:
-        return user
-    else:
+    user_for_deletion = await _get_user_by_id(user_id, session)
+    if not user_for_deletion:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"User with id {user_id} not found.",
         )
+    if not check_user_permissions(
+        target_user=user_for_deletion, current_user=current_user
+    ):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Forbidden.",
+        )
+    user = await _delete_user(user_id, session)
+    return user
 
 
 @user_router.get("/")
@@ -81,18 +89,26 @@ async def update_user(
             status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
             detail="At least one parameter for user update info must provided.",
         )
-    user = await _update_user(user_id, updated_user_params, session)
-    if user:
-        return user
-    elif user is False:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail=f"User with email {body.email} already exists.",
-        )
-    elif user is None:
+    user_for_update = await _get_user_by_id(user_id, session)
+    if not user_for_update:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"User with id {user_id} not found.",
+        )
+    if not check_user_permissions(
+        target_user=user_for_update, current_user=current_user
+    ):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Forbidden.",
+        )
+    user = await _update_user(user_id, updated_user_params, session)
+    if user:
+        return user
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"User with email {body.email} already exists.",
         )
